refactor(review): move cache helpers to api/, fix hook order, return type, cast comment, rename test (PR #563 R3)
