+++ src/entities/market-summary/__tests__/marketSummaryCache.test.ts
@@ -98,7 +98,7 @@ async function loadWithEnv(opts: { url?: string; token?: string }) {
     process.env.UPSTASH_REDIS_REST_URL = opts.url ?? '';
     process.env.UPSTASH_REDIS_REST_TOKEN = opts.token ?? '';
     vi.resetModules();
-    return import('../lib/marketSummaryCache');
+    return import('../api/marketSummaryCache');
 }
 
 describe('getCachedMarketSummary', () => {
