+++ src/entities/market-summary/__tests__/submitMarketBriefingAction.test.ts
@@ -6,11 +6,11 @@ import {
     type SubmitBriefingResult,
 } from '@y0ngha/siglens-core';
 import { isBot } from '@/shared/api/isBot';
-import { getCachedMarketSummary } from '../lib/marketSummaryCache';
+import { getCachedMarketSummary } from '../api/marketSummaryCache';
 
 vi.mock('server-only', () => ({}));
 
-vi.mock('../lib/marketSummaryCache', () => ({
+vi.mock('../api/marketSummaryCache', () => ({
     getCachedMarketSummary: vi.fn(),
 }));
