+++ src/entities/market-summary/api/briefingStaticCache.ts
