+++ src/entities/market-summary/api/marketSummaryStaticCache.ts
