+++ src/entities/sector-signal/__tests__/sectorSignalsCache.test.ts
@@ -52,7 +52,7 @@ async function loadWithEnv(opts: { url?: string; token?: string }) {
     process.env.UPSTASH_REDIS_REST_URL = opts.url ?? '';
     process.env.UPSTASH_REDIS_REST_TOKEN = opts.token ?? '';
     vi.resetModules();
-    return import('../lib/sectorSignalsCache');
+    return import('../api/sectorSignalsCache');
 }
 
 describe('getCachedSectorSignals', () => {
