+++ src/entities/sector-signal/api/sectorSignalsCache.ts
