+++ src/entities/sector-signal/api/sectorSignalsStaticCache.ts
