+++ src/shared/config/dashboard-tickers.ts
@@ -161,6 +161,10 @@ export const DEFAULT_DASHBOARD_TIMEFRAME: DashboardTimeframe = '1Day';
 export function isDashboardTimeframe(
     value: unknown
 ): value is DashboardTimeframe {
+    // The `as DashboardTimeframe` cast is safe here: the `.includes()` membership check
+    // is the actual runtime validation. The cast only satisfies TypeScript's strict
+    // `Array<T>.includes(searchElement: T)` parameter type — it does not bypass any
+    // runtime check or assume the input is already a valid timeframe.
     return DASHBOARD_TIMEFRAMES.includes(value as DashboardTimeframe);
 }
